update default payout perc and decay rate param types

+++ contracts/epochs/EpochsManager.sol
@@ -142,7 +142,7 @@ contract EpochsManager is IEpochsManager, Initializable, Ownable2StepUpgradeable
             uint128 baseLiveWinProb,
             uint128 expiredWinProb,
             uint256 ticketDuration,
-            uint16 decayRate
+            uint32 decayRate
         ) = _ticketingParameters.getTicketingParameters();
 
         uint256 nextEpochId = getNextEpochId();
